+++ TelegramBot.csproj
@@ -11,6 +11,8 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <PackageReference Include="Microsoft.Extensions.Configuration" Version="6.0.0" />
+    <PackageReference Include="Microsoft.Graph" Version="4.16.0" />
     <PackageReference Include="News.Net" Version="1.0.1" />
     <PackageReference Include="Telegram.Bot" Version="16.0.0" />
   </ItemGroup>
